Training seems to work now with gamma set to zero.  Need to check with non-zero gamma.
First pass at MPI parallelization.  It appears to be functional but need to check
that it converges still and runs faster of course.

There appear to be some bugs in the metric computations.  Doesn't affect
training but needs to be resolved for tensorboard debugging.

Note that there are experimental XLA flags on, these might be introducing
extra long compile times.  Turn off on line 13 of stochastic_reconfiguration.py